% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-html-tables.R
\name{col_group}
\alias{col_group}
\title{Create HTML table column group}
\usage{
col_group(ncl = NULL, widths = NULL)
}
\arguments{
\item{ncl}{Integer; Number of columns; derived from the length of the header row.}

\item{widths}{Numeric; Vector of column widths.}
}
\value{
Character; HTML \code{col_group} element
}
\description{
Creates \verb{<colgroup>} tag and child \verb{<col>} tags, for the purpose of setting
column widths. If \code{widths} is \code{NULL}, the widths are set equally to \verb{96\% / ncol}. If too few widths are provided, the table expands to fill the
HTML page.
}
\examples{
col_group(3, c(20, 30, 50)) |> cat(sep = '')

}
