% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs-maker.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert pairings list to tibble}
\usage{
convert_to_df(pairings, file_path = NULL)
}
\arguments{
\item{pairings}{A pairings list object returned by \code{create_pairs()}.}

\item{file_path}{Optional. A character string specifying a file path to save the output tibble as a CSV file. IF \code{NULL} (default), no file is saved.}
}
\value{
A tibble with the columns : \code{week} (Integer), \code{group} (Character), and \code{pairing} (Character).
}
\description{
A helper function to take a list of pairings returned by \code{create_pairs()}, and convert it into a tidy tibble
with one row per group per week.
Optionally, the tibble can be saved as a CSV file if \code{file_path} is specified.
}
\examples{
# Create a class list
roster <- LETTERS[1:10]

# Generate pairs
pairings <- create_pairs(roster)

# Convert to tibble
data_frame <- convert_to_df(pairings)

\dontrun{
# Save to CSV
data_frame <- convert_to_df(pairings, file_path = "Pairs.csv")
}

}
