% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs-maker.R
\name{create_pairs}
\alias{create_pairs}
\title{Create student pair-groups}
\usage{
create_pairs(pool, group_size = 2, record = NULL, population = 1000)
}
\arguments{
\item{pool}{Vector of student names}

\item{group_size}{Integer; Number of students per group}

\item{record}{List; List of pairings from previous weeks}

\item{population}{Integer; Number of simulated "organisms" to sample. I.e.
how long to spend searching for a better solution. Default is 1000.}
}
\value{
List of pairs groupings. Each list element is a character vector of
pairs.
}
\description{
Generate pair-groups from a list of student names, addressing the difficulty
of minimizing duplicate pairings in repeated groupings. #' Finding optimal
pairs can be complex, so this function uses a simulation approach inspired by
principles of evolutionary computing. It assesses the "fitness" of each
proposed group by penalizing duplicate pairings, thereby ensuring that the
fewest number of repeated pairs results in the highest fitness score.
Additionally, users can supply a record of previous pairings to inform the
fitness computation.
}
\examples{
#' ## for 1 week
num_students <- 8
create_pairs(LETTERS[1:num_students])

# ------------------------------------ #

## loop for n_weeks weeks
record <- NULL
n_weeks <- 5
for (i in 1:n_weeks){
  # to simulate varying numbers each week
  num_students <- sample(7:10, 1)

  record <- create_pairs(LETTERS[1:num_students],
                         record = record)
}

record

}
