% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-admissions-spreadsheet.R
\name{display_next}
\alias{display_next}
\title{Display next admissions data}
\usage{
display_next(d, i, n)
}
\arguments{
\item{d}{Admissions data frame of one row}

\item{i}{Integer; Current row number}

\item{n}{Integer; Total number of rows}
}
\description{
Handles user interaction. Returns user response.
}
