% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-admissions-spreadsheet.R
\name{filter_admissions}
\alias{filter_admissions}
\title{Filter admissions xlsx spreadsheat}
\usage{
filter_admissions(
  new_admissions_spreadsheet,
  reviewed_spreadsheet = NULL,
  decision_col = "team_action"
)
}
\arguments{
\item{new_admissions_spreadsheet}{Character; Path to admissions spreadsheet. Assumed to be the most recent spreadsheet}

\item{reviewed_spreadsheet}{Character; Path to previous will filter any that have had a note added already.}

\item{decision_col}{Character; Name of column used for recording decisions. Can be a vector of length two if each spreadsheet has decision col with a different name. Current default is \code{team_action}}
}
\value{
Data frame filtered for DSHSC applications. programme applications.
}
\description{
Reads the spreadsheet at the provided path. The spreadsheet should be stored
in your personal OneDrive space. The column names are changed to be
syntactic. The data are filtered to include only applications to "Data
Science for Health and Social Care" programmes (see
\href{programme_code_lookup}{}.) and only those with a \verb{Qualification Status Decision} equal to "\verb{(SB) Ready for Decision}".
}
\details{
No columns are removed, so the function should always work as intended, as
long as the \verb{Programme Code} column is correct, and is always called \verb{Programme Code}.
}
\examples{

filter_admissions('~/../OneDrive - University of Edinburgh/.../PG00x PG applicant communications report-online 01-Jan-00.xlsx')

}
