% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs-maker.R
\name{fitness_function}
\alias{fitness_function}
\title{Assess the "fitness" of a group}
\usage{
fitness_function(pairs, record)
}
\arguments{
\item{pairs}{List of character vectors, where each list element represent the pair two individuals that form a combined pair}

\item{record}{List of combined pairings}
}
\value{
Integer; the number of duplicate pairs
}
\description{
Counts the number of pairs in the current group that occur in the previous record of groups. Fewer matches means better fitness.
}
\examples{
\dontrun{

record <- list(
              c('AAA-BBB','WWW-XXX'),
              c('YYY-ZZZ', 'CCC-DDD')
          )

pairs <- list(
              c('AAA', 'BBB'),
              c('CCC', 'DDD'),
              c('EEE', 'FFF')
         )

fitness_function(pairs, record)
}

}
