% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-admissions-spreadsheet.R
\name{print.admissions}
\alias{print.admissions}
\title{Print method for admissions object}
\usage{
\method{print}{admissions}(x, all_cols = FALSE, ...)
}
\arguments{
\item{x}{Data frame of class \code{admissions}}

\item{all_cols}{Logical}

\item{...}{Ignored}
}
\description{
By default shows only a selected minimum of columns. To see all columns use
\code{all_cols = TRUE}. Column \code{team_action} is placed at the front of hte data frame for easy inspection, but is actually the final column of the admissions data set.
}
\keyword{internal}
