% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs-maker.R
\name{print.pairings}
\alias{print.pairings}
\title{Print method for pair record}
\usage{
\method{print}{pairings}(x, n = 1, ...)
}
\arguments{
\item{x}{Pairings; list}

\item{n}{Number of weeks to show}

\item{...}{Not used}
}
\value{
Invisibly returns x
}
\description{
Print method for pair record
}
\keyword{internal}
