% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programme-code-lookup.R
\docType{data}
\name{programme_code_lookup}
\alias{programme_code_lookup}
\title{Programme code lookup table}
\format{
\subsection{programme_code_lookup}{

A data frame, 7 rows by 7 columns.\tabular{llrl}{
   \strong{Variable} \tab \strong{Type} \tab \strong{N unique} \tab \strong{Example} \cr
   programme_code \tab character \tab 7 \tab PTMSCDSHSC1P \cr
   programme_description \tab character \tab 7 \tab Data Science for Health and Social Care.... \cr
   programme_name \tab character \tab 2 \tab Data Science for Health and Social Care \cr
   is_online \tab logical \tab 1 \tab \code{TRUE}, \code{FALSE} \cr
   qualification \tab character \tab 7 \tab MSc, PgDip, PgCert, etc. \cr
   length_years \tab integer \tab 5 \tab 1, 2, 3, 4, 6 \cr
   is_part_time \tab logical \tab 2 \tab \code{TRUE}, \code{FALSE} \cr
}

}
}
\usage{
programme_code_lookup
}
\description{
A table for mapping programme code to programme descriptions. Also contains
columns splitting the description into its  component parts.
}
\keyword{datasets}
