% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-admissions-spreadsheet.R
\name{progress_summary}
\alias{progress_summary}
\title{Called for side effect. Returns NULL.}
\usage{
progress_summary(d, decision_col)
}
\description{
Called for side effect. Returns NULL.
}
