% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs-maker.R
\name{sort_and_combine}
\alias{sort_and_combine}
\title{Collapse pairs into single string}
\usage{
sort_and_combine(pair)
}
\arguments{
\item{pair}{Vector of elements to combine}
}
\value{
Character vector, length 1
}
\description{
It is expected that \code{pair} will be a vector of characters (probably student
names), but will also work with a numeric vector.
}
\examples{

\dontrun{
sort_and_combine(c('ZZZ', 'AAA'))
# [1] "AAA~~ZZZ"
}

}
