% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-admissions-spreadsheet.R
\name{summary.admissions}
\alias{summary.admissions}
\title{Summary of admissions data}
\usage{
\method{summary}{admissions}(object, ...)
}
\description{
Count the number of decision statuses per qualification.
}
\keyword{internal}
