% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-html-tables.R
\name{table_body}
\alias{table_body}
\title{Create HTML table body}
\usage{
table_body(d, nrw, ncl)
}
\arguments{
\item{nrw}{Integer; Number of rows.}

\item{ncl}{Integer; Number of columns}
}
\value{
Character
}
\description{
Creates \verb{<tbody>} tag and child \verb{<tr>} and \verb{<td>} tags.
}
\examples{
table_body() |> cat(sep = '')

}
