% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-html-tables.R
\name{table_header}
\alias{table_header}
\title{Create HTML table header}
\usage{
table_header(d)
}
\arguments{
\item{col_names}{Character; Vector of column names.}
}
\value{
Character
}
\description{
Creates \verb{<thead>} tag and child \verb{<tr>} and \verb{<th>} tags, including the column
names.
s
}
\examples{
table_header(c('Criteria', 'Description', 'Weight')) |> cat()

}
