% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-funs.R
\name{update_from_github}
\alias{update_from_github}
\title{Pull changes from Github}
\usage{
update_from_github()
}
\value{
Invisibly returns names of any files added or changed on the Git
origin
}
\description{
For use in Noteable. Students can get the most recent notes into their
noteable environment when we release them. Files that have been changed
locally, that may cause a conflict, are backed up with a timestamp added to
their filename, e.g. \code{README.md} becomes
\code{README_local-backup_day-xxxxxx_time-zzzzzz.md}
}
