% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-html-tables.R
\name{update_table_style}
\alias{update_table_style}
\title{Update table CSS styles}
\usage{
update_table_style(...)
}
\arguments{
\item{...}{Named Character vector of CSS styles. The names are the selectors,
and the values are the CSS properties Position to insert the new styles}
}
\value{
Character; HTML \verb{<style>} element
}
\description{
Creates a \verb{<style>} tag with the provided CSS styles. Default styles are
included, and more styles can be added as a named vector. By default new
styles are added to the end of the default stylesheet. You can insert them at
a specific position using the \code{after} parameter.
}
\details{
Declarations added to existing selectors are appended to the existing
declarations for those selectors. Original declarations are not removed.
Rules for new selectors are appended to the end of the rules set.
}
\examples{

# `.row_highlight` already exists in the default stylesheet.
# The new declaration is simply appended next to the old.
# Selector that do not exist in the default stylesheet are added to the end of the stylesheet.
update_table_style(element = 'property: value;',
              .row_highlight = 'background-color: #000000;',
              `selector with spaces` = 'property: value;')

}
