% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usheR-package.R
\docType{package}
\name{usheR-package}
\alias{usheR}
\alias{usheR-package}
\title{usheR: Some helper functions for teaching}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: John Wilson \email{43386915+jmcvw@users.noreply.github.com}

}
\keyword{internal}
