% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance_function.R
\name{save_output}
\alias{save_output}
\title{Save or append a data frame to CSV}
\usage{
save_output(dataframe, file_path, append = FALSE)
}
\arguments{
\item{dataframe}{A data frame to be saved.}

\item{file_path}{A character string specifying the CSV file path.}

\item{append}{Logical. If \code{TRUE}, appends to an existing file if it exists. If
\code{FALSE}, the file is overwritten. Default is \code{FALSE}.}
}
\value{
Returns the (optionally saved) data frame.
}
\description{
This helper function saves a data frame to a specified \code{.csv} file path,
either by overwriting the file or appending to it. If \code{append = TRUE} and the
file already exists, the existing contents are read and combined with the new
data before being written back to the file. If \code{file_path} is \code{NULL}, the data
frame is returned without saving.
}
\examples{
# Load USJudgeRatings dataset and use as full class list
class_list <- data.frame(name = rownames(USJudgeRatings))

# print full class list to view row indices
print(class_list$name)

# Below is an example for saving
\dontrun{
# Week 1: take attendance
week1 <- take_attendance(full_class = class_list,
                         present_students = c(1, 2, 4),
                         file_path = "attendance.csv")  # saved using save_output()

# Week 2: take attendance and append to same file
week2 <- take_attendance(full_class = class_list,
                         present_students = c(1, 3, 4, 5, 7, 10, 15, 17),
                         file_path = "attendance.csv",
                         append = TRUE)

}
}
