% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-admissions-spreadsheet.R
\name{step_through_admissions}
\alias{step_through_admissions}
\title{Step through admissions one-by-one}
\usage{
step_through_admissions(admissions_data, reviewer)
}
\arguments{
\item{reviewer}{Character; Reviewer initials. Must be provided}

\item{d}{Admissions dataframe.}
}
\value{
Admissions dataframe with \code{team_action} column added.
}
\description{
Steps through an admissions spreadsheet one row at a time. The relevant
information for searching EUCLID is shown in the console, with the prompt
\verb{Action?}. The following actions can be taken:
\itemize{
\item \code{Enter}: The default action
\itemize{
\item A decision was made, and is recorded with reviewer's initials and date stamp.
}
\item \verb{#}: Skip
\itemize{
\item No decision was made. Can be returned to later.
}
\item \code{u}: Undo
\itemize{
\item Rewinds to the previous application.
}
\item \code{x}: Exit
\itemize{
\item Stops the review and returns an updated dataframe.
}
\item \code{q}: Quit
\itemize{
\item Same as \code{x}
}
}
}
\examples{
  step_through_admissions(admissions_data, 'jw')

}
